Removing top/bottom extra components and moving bib update to higher level
