+++ src/app/utils/bibDetailsUtils.jsx
@@ -1,6 +1,52 @@
 import React from 'react';
 import LibraryItem from './item';
 
+// `linkable` means that those values are links inside the app.
+// `selfLinkable` means that those values are external links and should be self-linked,
+// e.g. the prefLabel is the label and the URL is the id.
+const allFields = {
+  top: [
+    { label: 'Title', value: 'titleDisplay' },
+    { label: 'Author', value: 'creatorLiteral', linkable: true },
+    { label: 'Publication', value: 'publicationStatement' },
+    { label: 'Electronic Resource', value: 'React Component' },
+    {
+      label: 'Supplementary Content',
+      value: 'supplementaryContent',
+      selfLinkable: true,
+    },
+  ],
+  bottom: (bib) => [
+    {
+      label: 'Additional Authors',
+      value: 'contributorLiteral',
+      linkable: true,
+    },
+    { label: 'Found In', value: 'partOf' },
+    { label: 'Publication Date', value: 'serialPublicationDates' },
+    { label: 'Description', value: 'extent' },
+    { label: 'Donor/Sponsor', value: 'donor' },
+    { label: 'Series Statement', value: 'seriesStatement' },
+    { label: 'Uniform Title', value: 'uniformTitle' },
+    { label: 'Alternative Title', value: 'titleAlt' },
+    { label: 'Former Title', value: 'formerTitle' },
+    // if the subject heading API call failed for some reason,
+    bib.subjectHeadingData ? { label: 'Subject', value: 'subjectHeadingData' } : { label: 'Subject', value: 'subjectLiteral', linkable: true },
+    { label: 'Genre/Form', value: 'genreForm' },
+    { label: 'Notes', value: 'React Component' },
+    { label: 'Contents', value: 'tableOfContents' },
+    { label: 'Bibliography', value: '' },
+    { label: 'Call Number', value: 'identifier', identifier: 'bf:ShelfMark' },
+    { label: 'ISBN', value: 'identifier', identifier: 'bf:Isbn' },
+    { label: 'ISSN', value: 'identifier', identifier: 'bf:Issn' },
+    { label: 'LCCN', value: 'identifier', identifier: 'bf:Lccn' },
+    { label: 'OCLC', value: 'identifier', identifier: 'nypl:Oclc' },
+    { label: 'GPO', value: '' },
+    { label: 'Other Titles', value: '' },
+    { label: 'Owning Institutions', value: '' },
+  ],
+};
+
 const definitionItem = (value, index = 0) => {
   const link = (
     <a href={value.content} title={JSON.stringify(value.source, null, 2)}>
@@ -101,6 +147,7 @@ const getIdentifiers = (bib, detailsFields = []) => {
 };
 
 export {
+  allFields,
   annotatedMarcDetails,
   combineBibDetailsData,
   definitionItem,
